Renamed view to temp_view
