+++ lib/temp_view/bar.ts
