+++ lib/temp_view/chartlayermanager.ts
