+++ lib/temp_view/rule.ts
