+++ lib/temp_view/serieslabelstrip.ts
