+++ lib/temp_view/symbol.ts
