+++ lib/temp_view/tickstrip.ts
